// apps/api/src/ai-client/schema.d.ts
export interface components {
    schemas: {
// ...
        AssessmentResult: {
            /** Scores */
            scores: components["schemas"]["DriveScore"][];
            english: components["schemas"]["EnglishMetrics"];
            /** Interviewquestions */
            interviewQuestions: components["schemas"]["InterviewQuestion"][];
            candidateFeedback: components["schemas"]["CandidateFeedback"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}